JSE-15 Add constructors to FileBackedHTTPResource
https://issues.shibboleth.net/jira/browse/JSE-15

Add two new constructors to FileBackedHTTPResource that take a String rather than a Resource.  This avoids all sorts of ResourceLoader issues and allows the file name to specified without the garnish of file:://.

The parameter order is changed to maximise backwards compatibility.

Also add tests and change the (commented) example in the IdP distribution