JSE-25 Add RunnableResource and ScriptedRunnable
https://issues.shibboleth.net/jira/browse/JSE-25

Add a new Resource class which is but a thin veneer on the Spring
FileSystem one.  At initialization, when "lastModified", "exists"
and "getInputStream" are called a java.lang.Runnable is run.

This means that the runnable can be used to provoke any changes
(for instance running "svn update").

Add a ScriptedRunnable to allow scripting of this.  Plus tests.