IDP-1597 Change of language incomplete when switching locale
https://issues.shibboleth.net/jira/browse/IDP-1597

Add a new helper method which returns what the browser it set to,
but with Spring's idea of the "one true locale" prepended.

This allows the flexibility of falling back through the
browser alternative lamguages when Spring's "one true locale"
isn't known about.

It also makes for mimimal changes upstream.